Update routers folder name and add auth_router
