+++ src/routers/__init__.py
@@ -1,0 +1,2 @@
+from .auth import auth_router
+from .user import user_router
