+++ src/schemas/auth.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class TokenData(BaseModel):
+    access_token: str
+    token_type: str
